+++ src/routes/sis/courses/index.ts
@@ -1,0 +1,4 @@
+import { SISRouter } from "../router";
+import { CourseRouter } from "./router";
+
+SISRouter.use("/courses", CourseRouter);