+++ src/routes/sis/index.ts
@@ -4,4 +4,6 @@ import { SISRouter } from "./router";
 
 AppRouter.use("/sis", SISRouter);
 
+import "./courses";
+import "./programs";
 import "./students";