+++ src/routes/sis/programs/index.ts
@@ -1,0 +1,4 @@
+import { SISRouter } from "../router";
+import { ProgramRouter } from "./router";
+
+SISRouter.use("/programs", ProgramRouter);