+++ src/routes/users/router.ts
@@ -11,6 +11,7 @@ import {
     userAuthority,
     users,
 } from "../../user";
+import { isArray, isString } from "../../utils";
 
 export const UserRouter = Router();
 
@@ -20,6 +21,11 @@ UserRouter.post(
     auth.authorization((user) => user.role == UserRole.Administrator),
     async (req, res) => {
         const { email, password, role, info } = req.body;
+        if (!isArray<string>([email, password, role], isString) || !info)
+            return res
+                .status(400)
+                .json({ message: "Invalid or missin credentials" });
+
         const [status, user] = await users.create(email, password, role, info);
 
         switch (status) {
@@ -50,7 +56,7 @@ UserRouter.get("/me", auth.authenticate, ({ users = [] }: AuthRequest, res) => {
 
 UserRouter.post("/regen-pw", async (req: AuthRequest, res) => {
     const { credentialKey } = req.body;
-    if (!credentialKey || typeof credentialKey != "string")
+    if (!isString(credentialKey))
         return res.status(400).json({
             regenerated: false,
             message: "Invalid or missing credentials",