+++ src/utils/@types.ts
@@ -1,0 +1,1 @@
+export type ArrayElValidator<T> = (el: any) => el is T;